id type changed to long

+++ src/main/java/pl/futurecollars/invoicing/controller/invoice/InvoiceApi.java
@@ -21,7 +21,7 @@ public interface InvoiceApi {
 
     @ApiOperation(value = "Add new invoice to system")
     @PostMapping
-    int addInvoice(@RequestBody Invoice invoiceRq);
+    long addInvoice(@RequestBody Invoice invoiceRq);
 
     @ApiOperation(value = "Get single invoice by id")
     @GetMapping(value = "/{id}", produces = {"application/json;charset=UTF-8"})
